pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lbo-simulator"
version = "0.1.0"
description = "Leveraged Buyout Modeling & Capital Structure Optimizer"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "vanaprosto940-sudo"}
]
keywords = ["finance", "lbo", "private-equity", "modeling", "investment-banking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "pydantic>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "openpyxl>=3.1",
    "weasyprint>=59.0",
    "duckdb>=0.9",
    "plotly>=5.17",
    "streamlit>=1.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.1",
    "mypy>=1.5",
    "pre-commit>=3.4",
]
optimization = [
    "pulp>=2.7",
]

[project.scripts]
lbo-run = "lbo_simulator.scripts.run_lbo:main"
lbo-optimize = "lbo_simulator.scripts.optimize_structure:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/lbo_simulator --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
